% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_multi_tax_ids.R
\name{.expand_hits}
\alias{.expand_hits}
\title{Ensure you have run \code{.add_reference_subject_accession_versions} on the
\code{blastdbcmd_result} before running this function. If you don't, you will
hit an assertion error.}
\usage{
.expand_hits(blastdbcmd_result, blast_result)
}
\description{
NOTE: that we are "pretending" the original reference accessions only have a
single tax id after this function runs. That way things aren't counted more
than they should be, since they are now "expanded". TODO: need to double
check if the original rCRUX did this or not.
}
