% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{.validate_config_and_apply_default_params}
\alias{.validate_config_and_apply_default_params}
\title{Validate the raw config as read from the YAML file and apply defaults values
for any option that the user doesn't specify.}
\usage{
.validate_config_and_apply_default_params(config)
}
\description{
Validate the raw config as read from the YAML file and apply defaults values
for any option that the user doesn't specify.
}
