% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{.write_primers}
\alias{.write_primers}
\title{Writes enumerated primers to fasta files.}
\usage{
.write_primers(forward, reverse, chunk_count, to_directory)
}
\description{
The number is specified by the `chunk` count.
}
\details{
Returns the files names that were written.
}
