% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_multi_tax_ids.R
\name{expand_multi_tax_ids}
\alias{expand_multi_tax_ids}
\title{If \code{hits_with_multiple_taxonomy_ids} has zero rows, this function will
simply return that value back. That way you can join it with the single hits
and nothing will break, avoiding more conditional expressions.}
\usage{
expand_multi_tax_ids(
  blastdbcmd,
  hits_with_multiple_taxonomy_ids,
  blast_db_paths
)
}
\description{
If \code{hits_with_multiple_taxonomy_ids} has zero rows, this function will
simply return that value back. That way you can join it with the single hits
and nothing will break, avoiding more conditional expressions.
}
