% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_amplicon_blast_results.R
\name{parse_amplicon_blast_results}
\alias{parse_amplicon_blast_results}
\title{Parse amplicon blast results}
\usage{
parse_amplicon_blast_results(
  amplicon_blast_tsv_files,
  single_taxonomy_hits_outfile,
  multi_taxonomy_hits_outfile,
  parse_amplicon_blast_command,
  blastdbcmd,
  blast_db_paths,
  taxonomy_db_path
)
}
\arguments{
\item{ambiguous_run_limit}{\code{integer(1)}\cr
Sets the max allowed number of ambiguous bases in a row that a sequence
can have before it is thrown out.}
}
\description{
Parse amplicon blast results
}
\details{
We keep the longest alignment per subject accession. The "longest" is
based on the original \code{alignment_length} as reported by BLAST. After
this, the alignment is degapped and the alignment length is recalculated. So
be aware that the sequences are filtered based on original reported alignment
length, but then degapped alignment length is used going forward. This is the
behavior that matches the original rCRUX.
}
