% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipeline}
\alias{pipeline}
\title{Run the rCRUXMini pipeline}
\usage{
pipeline(config)
}
\value{
A list containing the pipeline results.
}
\description{
Note that the parameters in the config are _not_ validated by this function.
We assume you use the `new_config` creation function to create the config
object, which handles that.
}
