% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{rebind_global_logger}
\alias{rebind_global_logger}
\title{Rebind the global logger, e.g., to take into account values taken from the
config file.}
\usage{
rebind_global_logger(console_level, file_level = NULL, file_path = NULL)
}
\description{
This function should only be called _once_, and near the top of the
\code{pipeline} function after the directory in which you will be logging to
has been created. IGNORE THIS ADVICE AT YOUR OWN PERIL!
}
