% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_blastdbcmd.R
\name{run_blastdbcmd}
\alias{run_blastdbcmd}
\title{Run blastdbcmd}
\usage{
run_blastdbcmd(blastdbcmd, entry_batch_path, blast_db_path, outfmt)
}
\arguments{
\item{blastdbcmd}{todo}

\item{entry_batch_path}{todo}

\item{blast_db_path}{todo}

\item{outfmt}{todo}
}
\value{
An rcrux "result", either ok like \code{list(blast_db_path = blast_db_path, stdout = result$stdout)}, or an
  error, with a list containing the keys message, processx_object, condition.
  The last two may be NULL.
}
\description{
Run blastdbcmd
}
