% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_blastn.R
\name{run_blastn}
\alias{run_blastn}
\title{Wrapper for \code{SnailBLAST::crawl} to run \code{blastn} that specifies
standard rCRUX error handlers.}
\usage{
run_blastn(
  blast_executable_directory,
  query_paths,
  db_paths,
  outfmt_specifiers,
  extra_blast_arguments,
  use_long_names_in_parsed_result
)
}
\description{
Wrapper for \code{SnailBLAST::crawl} to run \code{blastn} that specifies
standard rCRUX error handlers.
}
