% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_command}
\alias{run_command}
\title{Like processx::run, except that it captures error and interrupt conditions and
rethrows them as \code{rcrux_mini_error} conditions.}
\usage{
run_command(command, args, ...)
}
\description{
\code{error_on_status = TRUE} is explicitly provided to \code{processx::run},
so don't include it in the dots.
}
