% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{test_file_non_empty}
\alias{test_file_non_empty}
\title{Checks if `path` is a file that exists and that it has a non-zero size.}
\usage{
test_file_non_empty(path)
}
\description{
If `path` refers to a directory this will fail.
}
